requirements.txt
spotipy>=2.23.0
python-dotenv>=1.0.0
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
pytest>=8.0.0
ruff>=0.3.0 